premium_chat/requirements.txt
# Django and Core Dependencies
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
python-decouple==3.8

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Authentication
django-allauth==0.57.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1

# API and Serialization
djangorestframework==3.14.0
django-filter==23.3

# AI Integration
g4f==0.1.9.9
requests==2.31.0
openai==1.3.7

# HTTP and Networking
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0

# Notion Integration
notion-client==2.0.0

# Image Processing
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==41.0.7

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Logging and Monitoring
django-extensions==3.2.3

# Security
django-csp==3.7
django-security==0.20.0
